Add Docker instructions for MongoDB
Currently the README specifies MongoDB as a dependency before starting
with the demo app and provides a link for installing it locally. In some
cases users may not want to install MongoDB on their local machines.

This adds information on how to use a Docker container for MongoDB and map
its port to the client.